requirements.txt
python-telegram-bot[asyncio]>=20.0
SQLAlchemy>=2.0
aiosqlite
python-dotenv
pytest
